% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrokenStick.R
\name{BrokenStick}
\alias{BrokenStick}
\title{Broken stick test for number of principal components}
\usage{
BrokenStick(n_dim, var_vec = NULL, tol = 0, plot = TRUE)
}
\arguments{
\item{n_dim}{the number of variables of the data (required).}

\item{var_vec}{the vector of explained variance of the PCA (optional).}

\item{tol}{a tolerance value that is used to check if the observed explained variance is larger than the expected under the Broken Stick model.}

\item{plot}{if a plot should be produced.}
}
\value{
Function will return the vector of explained variance for each PC. Will also return the number of PCs to keep and optionally a plot of the results if 'var_vec' is provided.
}
\description{
Function generates the expected distribution of variance across the principal components and compares it with the observed vector of explained variance (if provided).
}
