% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likFn.R
\name{loglik}
\alias{loglik}
\title{Probabilistic PCA likelihood function}
\usage{
loglik(L, sigma, N, d, g)
}
\arguments{
\item{L}{vector with length equal to the number of retained dimensions.}

\item{sigma}{variance. This is the average variance lost per discarded dimension.}

\item{N}{sample size. Number of observations.}

\item{d}{integer. Number of dimensions in the data.}

\item{g}{eigenvalues of the covariance matrix.}
}
\value{
returns the -1 * log-likelihood of the data.
}
\description{
Likelihood function for the standard probabilistic PCA or the phylogenetic probabilistic PCA. Note that one just need to change between the covariance of the data and the R matrix (multivariate Brownian Motion covariance matrix). The likelihood function only uses the eigenvalues of the covariance matrix (see parameter g).
}
\examples{
\donttest{
phy <- ratematrix::anoles$phy[[1]]
dt <- as.matrix( ratematrix::anoles$data[,1:3] )
L_start <- rgamma(n = 2, shape = 2, rate = 1)
sigma_start <- rgamma(n = 1, shape = 2, rate = 1)
R <- do3PCA:::fastR(x = dt, phy = phy)
g <- eigen(R)$values
loglik(L = L_start, sigma = sigma_start, N = nrow(dt), d = 2, g = g)
}

}
\references{
Revell, L. J. 2009. Size-Correction and Principal Components for Interspecific Comparative Studies. Evolution 63:3258–3268. doi: 10.1111/j.1558-5646.2009.00804.x

Revell, L. J. 2024. phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things). PeerJ 12:e16505. doi: 10.7717/peerj.16505

Tipping, M. E., and C. M. Bishop. 1999. Probabilistic Principal Component Analysis. Journal of the Royal Statistical Society Series B: Statistical Methodology 61(3):611–622. doi: 10.1111/1467-9868.00196

Zhang, A., Chan, K.L., Kwok, J.T., and D-Y Yeung. 2004. Proceedings of the AAAI Conference on Artificial Intelligence 19 Book One, Volume. https://aaai.org/papers/00372-aaai04-060-bayesian-inference-on-principal-component-analysis-using-reversible-jump-markov-chain-monte-carlo/
}
