% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPhyloPPCA.R
\name{mcmcPhyloPPCA}
\alias{mcmcPhyloPPCA}
\title{MCMC for Probabilistic Phylogenetic PCA}
\usage{
mcmcPhyloPPCA(
  phy,
  dt,
  ret_dim = 2,
  gamma_shape = 1,
  gamma_rate = 2,
  gen = 10^6,
  step_scale = 0.08,
  burn = 0.2
)
}
\arguments{
\item{phy}{phylogeny in "phylo" format.}

\item{dt}{a matrix with traits in columns and species values in rows. Rownames must match the tip labels of phylogeny.}

\item{ret_dim}{number of dimensions (PC axes) to be kept by the model.}

\item{gamma_shape}{The shape parameter for the gamma prior.}

\item{gamma_rate}{The rate parameter for the gamma prior.}

\item{gen}{The number generations for the MCMC.}

\item{step_scale}{The scaler for the mulplier proposal of the MCMC.}

\item{burn}{The proportion of the burn-in to the excluded before computing the average posterior parameter value.}
}
\value{
returns a list with "mean_scores", "percent_var_pcs", "var_loadings", and "stats". Also return "mcmc" which is the output from the "mcmc" R package.
}
\description{
Function to perform probabilistic phylogenetic PCA using Bayesian Markov-Chain Monte Carlo.
}
\details{
This function uses the "mcmc" R package to conduct a Metropolis-Hastings search. The output "mcmc" can be used for further analyses of the MCMC chain using the "mcmc" package.
}
\examples{
phy <- ratematrix::anoles$phy[[1]]
dt <- as.matrix( ratematrix::anoles$data[,1:3] )
## Do a VERY small chain to test:
mcmc3PCA <- mcmcPhyloPPCA(phy = phy, dt = dt, gen = 100)


}
\references{
Revell, L. J. 2009. Size-Correction and Principal Components for Interspecific Comparative Studies. Evolution 63:3258–3268. doi: 10.1111/j.1558-5646.2009.00804.x

Revell, L. J. 2024. phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other things). PeerJ 12:e16505. doi: 10.7717/peerj.16505

Tipping, M. E., and C. M. Bishop. 1999. Probabilistic Principal Component Analysis. Journal of the Royal Statistical Society Series B: Statistical Methodology 61(3):611–622. doi: 10.1111/1467-9868.00196

Zhang, A., Chan, K.L., Kwok, J.T., and D-Y Yeung. 2004. Proceedings of the AAAI Conference on Artificial Intelligence 19 Book One, Volume. https://aaai.org/papers/00372-aaai04-060-bayesian-inference-on-principal-component-analysis-using-reversible-jump-markov-chain-monte-carlo/
}
