% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaplaceApproximation.R
\name{modelChoice}
\alias{modelChoice}
\title{Automatic choice of number of principal components}
\usage{
modelChoice(x = NULL, phy = NULL, cov = NULL, N = NULL)
}
\arguments{
\item{x}{data. Columns are variables/features. Rows are observations/species.}

\item{phy}{phylogenetic tree. Number of tips must match number of rows in x.}

\item{cov}{a covariance matrix. Either the covariance of the data or an evolutionary ratematrix (R).}

\item{N}{number of observations/species in the tree.}
}
\value{
a list with the best fitting number of principal components (k) and a vector with the log probability of the data given the number of principal components (k).
}
\description{
Automatic choice of number of principal components
}
