% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbPCA.R
\name{ProbPCA}
\alias{ProbPCA}
\title{Probabilistic PCA}
\usage{
ProbPCA(x, ret_dim = 2, scale = TRUE)
}
\arguments{
\item{x}{a matrix with traits in columns and observations in rows.}

\item{ret_dim}{number of dimensions (PC axes) to be kept by the model.}

\item{scale}{if data should be rescaled to have zero mean and unit variance using the function scale(). Default is TRUE.}
}
\value{
returns a list.
}
\description{
Function to perform (non-phylogenetic) probabilistic PCA.
}
\details{
Optimization is performed using nloptr.
}
\examples{
dt <- as.matrix( ratematrix::anoles$data[,1:3] )
ppca <- ProbPCA(x = dt, ret_dim = 2)

}
\references{
Tipping, M. E., and C. M. Bishop. 1999. Probabilistic Principal Component Analysis. Journal of the Royal Statistical Society Series B: Statistical Methodology 61(3):611–622. doi: 10.1111/1467-9868.00196
}
