% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitModelChoice.R
\name{traitModelAIC}
\alias{traitModelAIC}
\title{Compute the MLE and AIC scores for a model of trait evolution}
\usage{
traitModelAIC(x, phy, model)
}
\arguments{
\item{x}{a dataset with traits in columns and species on the rows. Please make sure that dataset matches the phylogeny.}

\item{phy}{a phylogenetic tree with the same species as in x.}

\item{model}{the name of a model of trait evolution. The options are "BM", "lambda", "OU", or "EB".}
}
\value{
a list with the elements: model (the model name), lik (the log-likelihood), AIC (the Akaike Information Criterion), log_par (the log of the parameter value), nlopt_out (the output object from nloptr optimizer). If the model is "BM", then the output has R (the evolutionary R matrix).
}
\description{
Function will compute the parameter values and the AIC for a model of trait evolution. Models are estimated using a single parameter to model all traits in the multivariate data (x). The models are fit with the product of the likelihood of each independent trait fitted to a tree with the branch lengths transformed under the model M and a single model parameter p. The model parameter p represents an average of the process across all traits. The parameters for the models are: 1) lambda for Pagel's lambda, 2) alpha for OU (with theta as the root value), 3) b for the EB, and 4) sigma for Brownian motion. Please note that fitting BM does not require transformation of branch lengths.
}
